Fix menu lock-up
If you open one of the two currently present menues with either F2 or F3
and then open another one of these two menues again, it triggers a bug
that you can't exit these dialog menues any longer. This is a very crude
fix since I'm very new to rust. You might wish to refine this further
but for now it fixes #8